Put `ClientDiagnosticsPlugin` under `diagnostics` feature
And make it part of the `RepliconPlugins` group.
