+++ .github/workflows/main.yml
@@ -50,7 +50,7 @@ jobs:
         run: cargo clippy --benches --tests -- -D warnings
 
       - name: Rustdoc
-        run: cargo rustdoc -- -D warnings
+        run: cargo rustdoc --all-features -- -D warnings
 
   doctest:
     name: Doctest
@@ -66,7 +66,7 @@ jobs:
         uses: Swatinem/rust-cache@v2
 
       - name: Test doc
-        run: cargo test --doc
+        run: cargo test --all-features --doc
 
   test:
     name: Test
@@ -88,7 +88,7 @@ jobs:
         run: cargo install cargo-tarpaulin
 
       - name: Test
-        run: cargo tarpaulin --engine llvm --out lcov --exclude-files benches/* --exclude-files bevy_replicon_renet/*
+        run: cargo tarpaulin --all-features --engine llvm --out lcov --exclude-files benches/* --exclude-files bevy_replicon_renet/*
 
       - name: Upload code coverage results
         if: github.actor != 'dependabot[bot]'
