+++ src/lib.rs
@@ -453,10 +453,9 @@ pub mod prelude {
 
     pub use super::{
         client::{
-            diagnostics::{ClientDiagnosticsPlugin, ClientStats},
             events::ClientEventsPlugin,
             replicon_client::{RepliconClient, RepliconClientStatus},
-            ClientPlugin, ClientSet,
+            ClientPlugin, ClientSet, ClientStats,
         },
         core::{
             channels::{ChannelKind, RepliconChannel, RepliconChannels},
@@ -481,6 +480,9 @@ pub mod prelude {
         },
         RepliconPlugins,
     };
+
+    #[cfg(feature = "diagnostics")]
+    pub use super::client::diagnostics::ClientDiagnosticsPlugin;
 }
 
 pub use bincode;
@@ -493,12 +495,20 @@ pub struct RepliconPlugins;
 
 impl PluginGroup for RepliconPlugins {
     fn build(self) -> PluginGroupBuilder {
-        PluginGroupBuilder::start::<Self>()
+        let mut group = PluginGroupBuilder::start::<Self>();
+        group = group
             .add(RepliconCorePlugin)
             .add(ParentSyncPlugin)
             .add(ClientPlugin)
             .add(ServerPlugin::default())
             .add(ClientEventsPlugin)
-            .add(ServerEventsPlugin)
+            .add(ServerEventsPlugin);
+
+        #[cfg(feature = "diagnostics")]
+        {
+            group = group.add(ClientDiagnosticsPlugin);
+        }
+
+        group
     }
 }
